# alora-auth-api/docs/swagger.yaml
basePath: /
definitions:
...
  Company:
    properties:
      created_at:
        type: string
      domain:
        example: acme.com
        type: string
      domain_verified_at:
        type: string
      id:
        example: 5c6d7e8f-9a0b-1c2d-3e4f-5a6b7c8d9e0f
        type: string
      is_active:
        example: true
        type: boolean
      is_platform:
        example: false
        type: boolean
      max_seats:
        example: 100
        type: integer
      name:
        example: Acme Corp
        type: string
      subscription_status:
        enum:
        - TRIAL
        - ACTIVE
        - SUSPENDED
        - CANCELLED
        example: ACTIVE
        type: string
      updated_at:
        type: string
      user_count:
        example: 42
        type: integer
    type: object
...
  CreateCompanyRequest:
    properties:
      domain:
        example: acme.com
        maxLength: 253
        type: string
      is_active:
        example: true
        type: boolean
      max_seats:
        example: 100
        minimum: 1
        type: integer
      name:
        example: Acme Corp
        maxLength: 200
        minLength: 1
        type: string
      subscription_status:
        enum:
        - TRIAL
        - ACTIVE
        - SUSPENDED
        - CANCELLED
        example: ACTIVE
        type: string
    required:
    - name
    type: object
...
  CreateProductRequest:
    properties:
      accepts_api_clients:
        example: false
        type: boolean
      base_url:
        example: https://crm.acme.com
        maxLength: 2048
        type: string
      description:
        example: Customer relationships
        maxLength: 1000
        type: string
      initiate_login_uri:
        example: https://crm.acme.com/login/initiate
        maxLength: 2048
        type: string
      is_active:
        example: true
        type: boolean
      key:
        example: CRM
        maxLength: 40
        minLength: 2
        type: string
      name:
        example: Acme CRM
        maxLength: 200
        minLength: 1
        type: string
    required:
    - key
    - name
    type: object
  ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      reqId:
        example: 0b7c0f1e-6a19-4f3f-9a2d-1f9d0f3a6c21
        type: string
    type: object
...
  ProductRegistration:
    properties:
      accepts_api_clients:
        example: false
        type: boolean
      base_url:
        example: https://crm.acme.com
        type: string
      created_at:
        type: string
      description:
        type: string
      has_secret:
        example: true
        type: boolean
      id:
        example: 3f7c1b2e-5d4a-4c8b-9e10-2a6b7c8d9e0f
        type: string
      initiate_login_uri:
        example: https://crm.acme.com/login/initiate
        type: string
      is_active:
        example: true
        type: boolean
      key:
        example: CRM
        type: string
      name:
        example: Acme CRM
        type: string
      redirect_uris:
        example:
        - https://crm.acme.com/oidc/callback
        items:
          type: string
        type: array
      roles:
        example:
        - Admin
        - Editor
        - Viewer
        items:
          type: string
        type: array
      secret_rotated_at:
        type: string
      updated_at:
        type: string
    type: object
...
  ProductSecret:
    properties:
      client_id:
        example: 3f7c1b2e-5d4a-4c8b-9e10-2a6b7c8d9e0f
        type: string
      client_secret:
        example: acs_Rk9pX3VXYzV2cDFZa2dJd2Q0TGZaM2lPdkJjSmp0bFk
        type: string
      rotated_at:
        type: string
    type: object
...
  SetCompanyDomainRequest:
    properties:
      domain:
        example: acme.com
        maxLength: 253
        type: string
      verified:
        example: true
        type: boolean
    type: object
...
  Subscription:
    properties:
      ends_at:
        type: string
      id:
        example: 2b3c4d5e-6f70-8a9b-0c1d-2e3f4a5b6c7d
        type: string
      is_active:
        example: true
        type: boolean
      product_id:
        example: 3f7c1b2e-5d4a-4c8b-9e10-2a6b7c8d9e0f
        type: string
      product_key:
        example: CRM
        type: string
      product_name:
        example: Acme CRM
        type: string
      seat_limit:
        example: 50
        type: integer
      starts_at:
        type: string
    type: object
  SubscriptionRequest:
    properties:
      ends_at:
        type: string
      is_active:
        example: true
        type: boolean
      seat_limit:
        example: 50
        minimum: 1
        type: integer
    required:
    - is_active
    type: object
...
  UpdateCompanyRequest:
    properties:
      clear_max_seats:
        example: false
        type: boolean
      is_active:
        example: false
        type: boolean
      max_seats:
        example: 100
        minimum: 1
        type: integer
      name:
        example: Acme Corporation
        maxLength: 200
        minLength: 1
        type: string
      subscription_status:
        enum:
        - TRIAL
        - ACTIVE
        - SUSPENDED
        - CANCELLED
        example: SUSPENDED
        type: string
    type: object
...
info:
  contact:
    name: Alora Platform
    url: https://alora.io/support
  description: |-
    App Central: one sign-in for every product of the platform, and the OpenID Provider those products trust.

    ## Two kinds of token
    People sign in once at App Central — with a password, Google, or their company's SSO, as their company's login policy allows. That opens an App Central session (an HttpOnly cookie) and yields an App Central access token (audience `app-central`), the ONLY token the `/api` routes accept. It describes everything about App Central: `scope` (what the person may do here) and `products` (the keys of the apps they may open). A product never sees it: when a product is launched, its backend runs an ordinary OpenID Connect authorization-code flow with PKCE against `/oauth/authorize` and `/oauth/token`, and receives a token for itself alone (audience `product:<key>`, carrying only its own roles).

    ## Scopes
...

    ## Group managers
...

    ## Authorizing requests here
    Sign in with `POST /auth/login/password`, copy `access_token` from the response, then press **Authorize** and paste it. Every `/api` route reads the bearer token from the `Authorization` header only — never from a cookie — and re-checks its session against the database on every request.

    ## Companies
...

    ## Errors
...
  license:
    name: Proprietary
  termsOfService: https://alora.io/terms
  title: App Central API
  version: "2.0"
paths:
...
  /api/owner/companies:
    get:
      description: Every company, with its standing and member count. The platform
        company (`is_platform`) is the Owners' own.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Company'
            type: array
        "401":
          description: Missing, invalid or stale token
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Caller is not an Owner, or signed in more than twelve hours
            ago
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List companies
      tags:
      - owner
    post:
      consumes:
      - application/json
      description: Creates the company together with its Admins group and its default
        login policy (password and Google), in one statement. Invite its first Admin
        into the Admins group next.
      parameters:
      - description: The company
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Company'
        "400":
          description: Malformed body
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Caller is not an Owner, or signed in too long ago
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a company
      tags:
      - owner
  /api/owner/companies/{cid}:
    get:
      parameters:
      - description: Company id
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Company'
        "404":
          description: No such company
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a company
      tags:
      - owner
    patch:
      consumes:
      - application/json
      description: 'Send only what changes. `is_active: false` suspends the company:
        nobody in it can sign in, its App Central sessions stop working at once, and
        its products'' logins stop renewing. The platform company cannot be suspended.'
      parameters:
      - description: Company id
        in: path
        name: cid
        required: true
        type: string
      - description: Must repeat the company id
        in: header
        name: X-Alora-Target-Company
        required: true
        type: string
      - description: The changes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Company'
        "400":
          description: Malformed body, or the target header does not match
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: No such company
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: The platform company cannot be suspended
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a company
      tags:
      - owner
  /api/owner/companies/{cid}/domain:
    put:
      consumes:
      - application/json
      description: A verified domain is unique across companies. It is how the login
        page recognises the company's addresses, and what lets its SSO connections
        claim the domain.
      parameters:
      - description: Company id
        in: path
        name: cid
        required: true
        type: string
      - description: Must repeat the company id
        in: header
        name: X-Alora-Target-Company
        required: true
        type: string
      - description: The domain, or null
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SetCompanyDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Company'
        "409":
          description: Another company verified this domain
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set a company's email domain
      tags:
      - owner
...
  /api/owner/companies/{cid}/subscriptions:
    get:
      parameters:
      - description: Company id
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Subscription'
            type: array
        "404":
          description: No such company
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List a company's subscriptions
      tags:
      - owner
  /api/owner/companies/{cid}/subscriptions/{pid}:
    put:
      consumes:
      - application/json
      description: 'A subscription is switched off, never deleted. Switching it off
        ends the company''s access to the product: App Central stops offering it and
        every product login ends at its next refresh.'
      parameters:
      - description: Company id
        in: path
        name: cid
        required: true
        type: string
      - description: Product id
        in: path
        name: pid
        required: true
        type: string
      - description: Must repeat the company id
        in: header
        name: X-Alora-Target-Company
        required: true
        type: string
      - description: The subscription
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SubscriptionRequest'
      responses:
        "204":
          description: Set
        "404":
          description: No such company or product
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Subscribe a company to a product, or change it
      tags:
      - owner
...
  /api/owner/products:
    get:
      description: 'Every product App Central signs people in to, with its registration:
        redirect URIs, initiate_login_uri, role catalogue, and whether a client secret
        is set.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ProductRegistration'
            type: array
      security:
      - BearerAuth: []
      summary: List products
      tags:
      - owner
    post:
      consumes:
      - application/json
      description: 'The key is permanent: every token for the product carries the
        audience `product:<key>`. Then set its redirect URIs and roles, and issue
        its client secret.'
      parameters:
      - description: The product
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ProductRegistration'
        "400":
          description: Malformed body, a bad key or URL
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: The key is taken
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a product
      tags:
      - owner
  /api/owner/products/{pid}:
    get:
      parameters:
      - description: Product id
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ProductRegistration'
        "404":
          description: No such product
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a product
      tags:
      - owner
    patch:
      consumes:
      - application/json
      description: The complete settable state, the key excepted. An inactive product
        signs nobody in and renews no login. `accepts_api_clients` lets companies'
        API clients get tokens for the product (off until switched on); leave it out
        to keep it as it is.
      parameters:
      - description: Product id
        in: path
        name: pid
        required: true
        type: string
      - description: The product
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ProductRegistration'
        "404":
          description: No such product
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - owner
  /api/owner/products/{pid}/client-secret:
    post:
      description: A new secret, shown in this response only and stored as a hash.
        The previous one stops working at once, so deploy this one to the product's
        backend straight away.
      parameters:
      - description: Product id
        in: path
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ProductSecret'
        "404":
          description: No such product
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Issue a product's client secret
      tags:
      - owner
  /api/owner/products/{pid}/redirect-uris:
    put:
      consumes:
      - application/json
      description: The COMPLETE list. An authorization code is delivered only to one
        of these, matched byte for byte — no prefixes, no wildcards.
      parameters:
      - description: Product id
        in: path
        name: pid
        required: true
        type: string
      - description: The redirect URIs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ProductRedirectURIsRequest'
      responses:
        "204":
          description: Set
        "400":
          description: A URI that is not absolute https, or carries a fragment
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: No such product
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set a product's redirect URIs
      tags:
      - owner
  /api/owner/products/{pid}/roles:
    put:
      consumes:
      - application/json
      description: The COMPLETE catalogue. Grants may name only these roles, and a
        role still granted cannot be removed.
      parameters:
      - description: Product id
        in: path
        name: pid
        required: true
        type: string
      - description: The roles
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ProductRolesRequest'
      responses:
        "204":
          description: Set
        "400":
          description: A malformed role name
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: No such product
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: A role being removed is still granted
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set a product's role catalogue
      tags:
      - owner
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'An App Central access token from /auth/login/password or /auth/central/refresh.
      Enter it as: Bearer &lt;token&gt;'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...